move TryIntoError into a dedicated file
